+++ src/aio.c
@@ -198,6 +198,27 @@ static void raio_complete(void *arg) {
 
 }
 
+static void raio_complete_close(void *arg) {
+
+  nano_aio *raio = (nano_aio *) arg;
+  const int res = nng_aio_result(raio->aio);
+  if (res == 0) {
+    nng_msg *msg = nng_aio_get_msg(raio->aio);
+    raio->data = msg;
+    nng_pipe p = nng_msg_get_pipe(msg);
+    nng_pipe_close(p);
+  }
+
+#if NNG_MAJOR_VERSION == 1 && NNG_MINOR_VERSION < 6
+  nng_mtx_lock(shr_mtx);
+  raio->result = res - !res;
+  nng_mtx_unlock(shr_mtx);
+#else
+  raio->result = res - !res;
+#endif
+
+}
+
 static void raio_complete_signal(void *arg) {
 
   nano_cv_aio *cv_aio = (nano_cv_aio *) arg;
@@ -1483,7 +1504,7 @@ SEXP rnng_ncurl_session_close(SEXP session) {
 
 // request ---------------------------------------------------------------------
 
-SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP clo) {
+SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeout, SEXP autoclose, SEXP clo) {
 
   if (R_ExternalPtrTag(con) != nano_ContextSymbol)
     Rf_error("'context' is not a valid Context");
@@ -1492,6 +1513,7 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
 
   int xc;
   const nng_duration dur = timeout == R_NilValue ? NNG_DURATION_DEFAULT : (nng_duration) Rf_asInteger(timeout);
+  const int ac = LOGICAL(autoclose)[0];
 
   SEXP sendaio, aio, env, fun;
   nano_buf buf;
@@ -1519,11 +1541,10 @@ SEXP rnng_request(SEXP con, SEXP data, SEXP sendmode, SEXP recvmode, SEXP timeou
   NANO_FREE(buf);
 
   nano_aio *raio = R_Calloc(1, nano_aio);
-
   raio->type = RECVAIO;
   raio->mode = nano_matcharg(recvmode);
 
-  if ((xc = nng_aio_alloc(&raio->aio, raio_complete, raio))) {
+  if ((xc = nng_aio_alloc(&raio->aio, ac ? raio_complete_close : raio_complete, raio))) {
     R_Free(raio);
     nng_aio_free(saio->aio);
     R_Free(saio);